fix: versioned updates should also be resilient to concurrent transactions
