+++ xoto3/__about__.py
@@ -1,4 +1,4 @@
 """xoto3"""
-__version__ = "1.10.1"
+__version__ = "1.10.2"
 __author__ = "Peter Gaultney"
 __author_email__ = "pgaultney@xoi.io"
